Add database

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -9,7 +9,6 @@ import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.film.InMemoryFilmStorage;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
@@ -46,7 +45,6 @@ public class FilmController {
             log.info("Новый фильм успешно добавлен");
             return film;
         } catch (ValidationException exception) {
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new ValidationException(exception.getMessage());
         }
     }
@@ -55,15 +53,16 @@ public class FilmController {
     public Film update(@RequestBody @Valid Film film)  {
         log.info("Получен запрос к эндпоинту: PUT: /films");
         try {
+            if(film.getId() < 0){
+                throw new NullPointerException("Некорректно задан id фильма");
+            }
             Check.checkFilm(film);
-            film = filmService.update(film);
+            Film film1 = filmService.update(film);
             log.info("Фильм успешно обновлен");
-            return film;
+            return film1;
         } catch (ValidationException exception) {
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new ValidationException(exception.getMessage());
         } catch (NullPointerException exception) {
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new NullPointerException(exception.getMessage());
         }
     }
@@ -77,32 +76,35 @@ public class FilmController {
     public Film getFilmById (@PathVariable("id") Integer id) {
         log.info("Получен запрос к эндпоинту: GET: /films/{id}");
         try {
+            if(id <= 0) {
+                throw new NullPointerException("Некорректно задан id фильма");
+            }
             return filmService.getFilmById(id);
         } catch (NullPointerException exception){
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new NullPointerException(exception.getMessage());
         }
 
     }
 
     @PutMapping ("/{id}/like/{userId}")
-    public Film addLike (@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void addLike (@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
         log.info("Получен запрос к эндпоинту: PUT: /films/{id}/like/{userId}");
         try {
-            return filmService.addLike(id, userId);
+            filmService.addLike(id, userId);
         } catch (NullPointerException exception){
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new NullPointerException(exception.getMessage());
         }
     }
 
     @DeleteMapping ("/{id}/like/{userId}")
-    public Film removeLike(@PathVariable Integer id, @PathVariable Integer userId) {
+    public boolean removeLike(@PathVariable Integer id, @PathVariable Integer userId) {
         log.info("Получен запрос к эндпоинту: DELETE: /films/{id}/like/{userId}");
         try {
+            if(id <= 0 || userId <= 0) {
+                throw new NullPointerException("Некорректно задан id фильма или userId");
+            }
             return filmService.removeLike(id, userId);
         } catch (NullPointerException exception){
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new NullPointerException(exception.getMessage());
         }
     }
@@ -119,11 +121,10 @@ public class FilmController {
     }
 
     @DeleteMapping ("/{id}")
-    public void removeFilmById(@PathVariable Integer id) {
+    public boolean removeFilmById(@PathVariable Integer id) {
         try{
-            filmService.removeFilmById(id);
+            return filmService.removeFilmById(id);
         } catch (NullPointerException exception){
-            log.info("Возникла ошибка: " + exception.getMessage());
             throw new NullPointerException(exception.getMessage());
         }
     }
